pyproject.toml
[tool.poetry]
name = "smart-store"
version = "0.1.0"
description = ""
authors = ["kimdooh <johnkdo2020@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
Django = "<4.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
django-phonenumber-field = "^6.1.0"
phonenumbers = "^8.12.47"
django-extensions = "^3.1.5"
notebook = "^6.4.11"
supervisor = "^4.2.4"
gunicorn = "^20.1.0"
pandas = "^1.4.2"
numpy = "^1.22.3"
beautifulsoup4 = "^4.11.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
